Flatpak support
Add a 'koji-containerbuild flatpak-build' subcommand that triggers
building a flatpak rather than a Docker container. Most of the code
and options are shared with koji-containerbuild, but a --module
option is added, and since a new koji/osbs can be assumed:

 * The --git-branch option is required, as required by recent OSBS
 * source archive uploads are forbidden (did this ever work?),
   and the --no-progress option is removed.
 * The --release option is removed

In the builder plugin, the flatpak and module options are passed to osbs.